UTFPR_Grade_Builder/requirements.txt
PySide6>=6.7
playwright>=1.50
Pillow>=10.0
python-dotenv>=1.0
pytest>=8.0
ruff>=0.9
